The lexers in this project now use the new str_buf_t to store strings rather than APR arrays.  This should definitely be more efficient since we will not have the overhead of a function call for each character stored.
It will probably make sense to further expand this API and have several
different types of string buffers, such as one that cannot grow.  When parsing
a template it probably is smarter to just fill the buffer and then make that a
text segment rather than to keep growing the buffer.