Changed the way format functions work.  There is now a new function called jxtl_template_register_format().  This function registers a named format with a callback function.
This change allows more flexibility with formatting, since it does not have to
all be done from the same function.  Note:  the prototype for a format function
callback has not changed, it still takes the format name as its second
parameter.   Keeping the prototype the same allows very trivial formatting, i.e.
uppercase/lowercase, to be combined into one function if the user so chooses.