We now use expat directly to convert XML to JSON.
This should help improve performance slightly since it does
not require the entire XML document to be read into memory first.

I also fixed many compile warnings.  We now just use chars.  The only
reason we need unsigned is when we are actually looking at the value
of a char, which is not very often.

Also, XML containing mixed content now generates an errror.