Better parsing of path expressions.
The parser would error if you did something like {{#if ! expr }}.
The problem was that were expecting the path expression to have no
spaces in it.  Going forward, this requirement is unrealistic.  There
is now a separate section in lex file for the path expression.

Now newlines and spaces can be in there with no problem.  I also changed
test 9 to cover this case.