A slight change to how JSON is dumped.
The characters are U+2028 (line separator) and U+2029 (paragraph separator)
are checked for when dumping JSON.  If these are found, we output them
as their Unicode escape sequence rather than literally.  There are issues
if you try to embed JSON in a web page and output these characters
literally.  Browsers treat them as newlines, which causes the a JSON
parse error